pyproject.toml
[project]
name = "prompt-analyzer"
version = "0.1.0"
description = "A tool for analyzing and refining prompts using multiple LLMs"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",
    "textual >= 0.89.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
select = ["E", "F", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py39"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "black>=24.0.0", "ruff>=0.1.0"]
